+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-\-
+### Added
+- Lost Tribe tokens.
 
 ## [0.6.0] - 2022-04-26
 ### Added